Add sample SSH config file, link to docs
Since we're in the middle of removing WATonomous server specific
documentation in joint-bvs-ws, we need to make sure that our docs are
sufficient for users to get started. One thing that I found is lacking
is mention of SSH configuration files. I don't want to write a whole
guide on it, but I think linking good documentation as well as a sample
configuration is minimal yet enough to get most started.